chore: refactor selectors into separate directory
