+++ src/components/AssetSelection/components/AssetChainDropdown/AssetChainDropdown.tsx
@@ -12,9 +12,9 @@ import type { AssetId } from '@shapeshiftoss/caip'
 import { memo, useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { useWallet } from 'hooks/useWallet/useWallet'
-import { isAssetSupportedByWallet } from 'state/slices/portfolioSlice/utils'
-import { selectRelatedAssetIdsInclusiveSorted } from 'state/slices/related-assets-selectors'
 import { selectChainDisplayNameByAssetId } from 'state/selectors'
+import { selectRelatedAssetIdsInclusiveSorted } from 'state/selectors/related-assets-selectors'
+import { isAssetSupportedByWallet } from 'state/slices/portfolioSlice/utils'
 import { useAppSelector } from 'state/store'
 
 import { AssetRowLoading } from '../AssetRowLoading'
