+++ src/components/Layout/Header/GlobalSearch/AssetResults/AssetResult.tsx
@@ -6,13 +6,13 @@ import { AssetIcon } from 'components/AssetIcon'
 import { RawText } from 'components/Text'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import { middleEllipsis } from 'lib/utils'
-import type { GlobalSearchResult } from 'state/slices/search-selectors'
-import { GlobalSearchResultType } from 'state/slices/search-selectors'
 import {
   selectAssetById,
   selectPortfolioCryptoPrecisionBalanceByFilter,
   selectPortfolioUserCurrencyBalanceByAssetId,
 } from 'state/selectors'
+import type { GlobalSearchResult } from 'state/selectors/search-selectors'
+import { GlobalSearchResultType } from 'state/selectors/search-selectors'
 import { useAppSelector } from 'state/store'
 
 import { ResultButton } from '../ResultButton'
