+++ src/components/Layout/Header/GlobalSearch/TxResults/TxResult.tsx
@@ -1,8 +1,7 @@
 import { forwardRef } from '@chakra-ui/react'
 import { useCallback } from 'react'
 import { TransactionRow } from 'components/TransactionHistoryRows/TransactionRow'
-import type { GlobalSearchResult } from 'state/slices/search-selectors'
-import { GlobalSearchResultType } from 'state/slices/search-selectors'
+import { type GlobalSearchResult, GlobalSearchResultType } from 'state/selectors/search-selectors'
 import type { TxId } from 'state/slices/txHistorySlice/txHistorySlice'
 
 type TxResultProps = {
