+++ src/hooks/useProfileAvatar/useProfileAvatar.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 import { makeBlockiesUrl } from 'lib/blockies/makeBlockiesUrl'
 import { selectSelectedNftAvatarUrl } from 'state/apis/nft/selectors'
-import { selectWalletId } from 'state/slices/common-selectors'
+import { selectWalletId } from 'state/selectors/common-selectors'
 import { useAppSelector } from 'state/store'
 
 export const useProfileAvatar = () => {
