+++ src/state/selectors/search-selectors.ts
@@ -1,10 +1,9 @@
 import type { AssetId } from '@shapeshiftoss/caip'
 import type { Asset } from '@shapeshiftoss/types'
-import { createDeepEqualOutputSelector } from 'state/selector-utils'
-
-import { selectSearchQueryFromFilter } from '../selectors'
-import { selectAssetsBySearchQuery, selectTxsByQuery } from '../selectors'
-import type { TxId } from './txHistorySlice/txHistorySlice'
+import { createDeepEqualOutputSelector, selectSearchQueryFromFilter } from 'state/selector-utils'
+import { selectAssetsBySearchQuery } from 'state/slices/assetsSlice/selectors'
+import { selectTxsByQuery } from 'state/slices/txHistorySlice/selectors'
+import type { TxId } from 'state/slices/txHistorySlice/txHistorySlice'
 
 export enum GlobalSearchResultType {
   Asset = 'Asset',
