+++ src/state/slices/opportunitiesSlice/resolvers/cosmosSdk/index.ts
@@ -4,8 +4,8 @@ import { isFulfilled, isRejected, isSome } from 'lib/utils'
 import { accountIdToFeeAssetId } from 'lib/utils/accounts'
 import { assertGetCosmosSdkChainAdapter } from 'lib/utils/cosmosSdk'
 import type { ReduxState } from 'state/reducer'
+import { selectWalletAccountIds } from 'state/selectors/common-selectors'
 import { selectAssetById } from 'state/slices/assetsSlice/selectors'
-import { selectWalletAccountIds } from 'state/slices/common-selectors'
 import { selectMarketDataByAssetIdUserCurrency } from 'state/slices/marketDataSlice/selectors'
 
 import type {
