+++ src/state/slices/opportunitiesSlice/resolvers/foxy/index.ts
@@ -4,8 +4,8 @@ import dayjs from 'dayjs'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { foxyApi } from 'state/apis/foxy/foxyApi'
 import { getFoxyApi } from 'state/apis/foxy/foxyApiSingleton'
+import { selectPortfolioCryptoBalanceBaseUnitByFilter } from 'state/selectors/common-selectors'
 import { selectAssetById } from 'state/slices/assetsSlice/selectors'
-import { selectPortfolioCryptoBalanceBaseUnitByFilter } from 'state/slices/common-selectors'
 import { selectMarketDataByAssetIdUserCurrency } from 'state/slices/marketDataSlice/selectors'
 import { selectBIP44ParamsByAccountId } from 'state/slices/portfolioSlice/selectors'
