+++ src/state/slices/opportunitiesSlice/resolvers/uniV2/index.ts
@@ -12,8 +12,8 @@ import { toBaseUnit } from 'lib/math'
 import { selectZapperFulfilled } from 'state/apis/zapper/selectors'
 import { zapperApi } from 'state/apis/zapper/zapperApi'
 import type { ReduxState } from 'state/reducer'
+import { selectPortfolioAccountBalancesBaseUnit } from 'state/selectors/common-selectors'
 import type { AssetsState } from 'state/slices/assetsSlice/assetsSlice'
-import { selectPortfolioAccountBalancesBaseUnit } from 'state/slices/common-selectors'
 import { marketData } from 'state/slices/marketDataSlice/marketDataSlice'
 import { selectMarketDataByAssetIdUserCurrency } from 'state/slices/marketDataSlice/selectors'
 import type { PortfolioAccountBalancesById } from 'state/slices/portfolioSlice/portfolioSliceCommon'
